+++ java/avro-converter/pom.xml
@@ -1,7 +1,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>com.microsoft</groupId>
+  <groupId>com.microsoft.azure.schemaregistry.kafka.connect.avro</groupId>
   <artifactId>SchemaRegistryAvroConverter</artifactId>
   <packaging>jar</packaging>
   <version>1.0.0-beta.1</version>
@@ -21,12 +21,6 @@
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
       <version>3.8.1</version>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.junit.jupiter</groupId>
-      <artifactId>junit-jupiter-api</artifactId>
-      <version>5.9.2</version>
       <scope>test</scope>
     </dependency>
     <dependency>
@@ -82,5 +76,4 @@
         </plugin>
     </plugins>
   </build>
-
 </project>
